Hold Notice Message Enhancements.
Notification added to the success message after hold is placed. Shows &
highlights the phone number, email, or address to be notified.
There is an edit icon & link on the information, if the user is allowed
to update profile.
If no notification preference is specified, no notification method is
mentioned, just that the user will be notified. In this case a link is
presented to set notification preferences if the library has allowed
Notification preferences to be changed.