PK-784 ratings/review in reading history
http://marmot.myjetbrains.com/youtrack/issue/PK-784

Work id now properly populated to rating html elements on reading
history.  Server-side checking in place to ensure an id and a valid
rating are sent. (This prevents updating the first rating in the
database  when no id is given). Also added a check for same value or
rating given by user, as the update method returns failure when the
value doesn't change.

Changed several pop-up button using span tags to use button tags so that
navigating with the keyboard is possible in the pop-up windows.

#PK-784