PK-783 opt out reviews after ratings
http://marmot.myjetbrains.com/youtrack/issue/PK-783

add a prompt form on the server side.
if the user has already opted out , the form won't be shown.
account profile now has a My Preferences Section containing the opt out
setting so that users may later opt back in.

I moved removal of the memcache copy into the User object's update()
method as well as the $_SESSION updating, so that any calls to update
the user in the database will clear the cache & set the $_SESSION.
Hopefully this will make future coding related to user profile/account
cleaner.

#pk-783