PK-837 All tabs on a tabbed List Widget display the same results.
http://marmot.myjetbrains.com/youtrack/issue/PK-837

Create Widget link on Search Results page has validation requiring a
name for the widget or tab to prevent nameless widgets.  The form's
select list is alphabetically now so that the desired existing widget is
more easily found. The form also received some basic styling
improvements.

Create Widget on the server side now has a check that a value was given
for source ID and source.

reference to $().tabs() was commented out. browser console was giving
error messages. Doesn't seem to have an effect.

#PK-837