PK-820 Screen Scraping to Book an Item
http://marmot.myjetbrains.com/youtrack/issue/PK-820

The Screen Scraping for Bookings Module, as well as, the AJAX handlers,
both client & server side.

The user form has calendar date pickers, as well as autocomplete
suggestions for time.  Client-side validation is in place for all
fields. Server-side will attempt a booking as long as it has a good
record id and a start date.

If the booking recieves an error message from the webpac interface, that
is sent back to the user along with the form so that the user may
attempt to reschedule.

The cURL operations have been packaged into two functions & two object
properties in clean manner that hopefully can be used in other screen
scraping classes.

#PK-820