PK-819 Availability Calendar
http://marmot.myjetbrains.com/youtrack/issue/PK-819

Screenscapes the availability table from the sierra webpac, showing the
available hours for the next 30 days. This is done via an ajax call so
that the user can start using the form, with out waiting for the
calendar information.  If the ajax call successfully returns html, the
Show Hourly Calendar is then presented to the user.  The button will
toggle the display of the calendar.

The calendar table was simplified so that available hours are shown with
a green background, unavailable or closed hours are grayed. #PK-819