PK-815 Show Book Material Button based on item's bookable status
http://marmot.myjetbrains.com/youtrack/issue/PK-815

The bookable status was added to both the sierra & millennium status
loaders. Through the ajax call for holdings information on the full
record view, this information is transmitted to the user's browser,
which will then show the Book Material button if the ShowBookMaterial
xml element (in the ajax response) was 1.

Currently that switch is turned on if there are any bookable copies in
the holdings. #PK-815