PK-913 Implement Account Linking Blocks
http://marmot.myjetbrains.com/youtrack/issue/PK-913

Blocks are implemented when getting linked accounts for user and when a
new account  is being linked.

success/failure return added to the addLinkedUser method so that a
failure message can be given to the user.  Message is just that account
linking failed, even in the case of blocking.  Small javascript
adjustment so that the messages would be displayed correctly to user.
#PK-913