PK-915 Renew Selected works will linked accounts
http://marmot.myjetbrains.com/youtrack/issue/PK-915

smarty template modified to pass back user id for the checked out item,
and record id, in addition to the renewal Indicator.  The AJAX processor
calls the appropriate renewItem method via the user object. The
processor collects success & failure message as before to return to
user.

The millennium checkouts class was updated to use the screen-scraping
class. #PK-915