PK-819 Booking Calendar Fixes
http://marmot.myjetbrains.com/youtrack/issue/PK-819

The "ils:" prefix caused the booking calendar not to display. The prefix
is removed from the ajax url, and is removed from the recordId when
received from the ajax call.  Also the user object is used to get the
catalog driver, in order to call the bookings driver.  The Prefix is
also removed from recordId for the Booking action. #PK-819