PK-931 Fines page handling multiple accounts
http://marmot.myjetbrains.com/youtrack/issue/PK-931

The fines of each linked account will be displayed with a table per
account, underneath a header of the account name. (Account Name format
is the same as shown in account settings.)  Each table has a footer now
displaying the fine totals.

Alert at the top of the page is for the primary account only, as well as
the pay fine button at the bottom, if & when it appears. #PK-931