PK-932 updating Account Settings with linked accounts
http://marmot.myjetbrains.com/youtrack/issue/PK-932

A select form allows the user to switch to each of their linked
accounts, allowing the user to update settings for that account. This
form was made as it's own template so that it can easily be added into
other MyAccount pages.

The settings for the home library of the chosen account are put in
effect when that account is selected, so that library's settings take
effect regardless of the interface the main account is using.

The Add Account button is disabled when the account being displayed is
not the main account.  Instead there is a short notice telling the user
to log in with that account to add other accounts.

The Catalog Options & User Preferences Sections were combined into one
User Preferences section.  The updating functions were also combined
into one.

The user object deletePatronProfileCache() was updated to properly match
the cache keys being stored now.  This function should be preferred over
delete profile cache functions in catalog drivers.  Some of those
functions were rewritten to use deletePatronProfileCache() #Pk-932