PK-925 Additional Reading History Fix
http://marmot.myjetbrains.com/youtrack/issue/PK-925

Fix one: When patronId is passed to readingHistory.php as an empty
value, the patronId would be set to the empty value. Now an empty value
will default to the logged-in/primary user as intended.

Fix two:  When the primary user account has no linked accounts, the
readingHistory.php script did not set the template variable
$selectedUser. However this variable is needed for the reading history
actions, including OptIn. (Actions should have defaulted to the primary
user, but for problem addressed by fix one.) #PK-925