PK-947 Pickup Hold Location not setting
http://marmot.myjetbrains.com/youtrack/issue/PK-947

Pickup location code gets padded with spaces prior to be submit to the
webpac interface. Previously the spaces were pre-url-encoded to pluses.

The MyAccount/Holds display was also not correctly parsing the pickup
location from the select option tags in the screen scraping. The
location coding padding is here also, except that the selected option
value is padded with spaces and not plus signs. ( This is the only
option value that does not have the pluses.) #PK-947