PK-972  Test web server setup to reflect planned setup for production
http://marmot.myjetbrains.com/youtrack/issue/PK-972

The two httpd configuration files allow the appropriate virtual host to
handle web requests based on URL.  This includes the secure servers
using the appropriate security certificates.  This now enables aspencat
logging to be stored  in the aspencat log files! (both error and access)

Each configuration file will be linked to from the /etc/httpd/conf.d
directory.

Important:  For the Marmot test server, the Aspencat configuration must
be loaded first for the name-based virtual hosts to work correctly. This
is because the server alias *.marmot.org would include all the test
aspencat sites (*.aspencat.marmot.org). When the aspencat configuration
is loaded first, it's alias gets checked first. This is neccessary on
the test server. This order *shouldn't* be  necessary on the production
server. #PK-972