PK-1035 Auto log out for Item Level Holds
http://marmot.myjetbrains.com/youtrack/issue/PK-1035

When Auto log out is selected on the place hold form, and the hold
requires an item level hold, the AJAX.php placeHold() method would
prematurely log the user out (before the user is given the chance to
choose the item to place a hold on).

Now if an item level hold is needed the user is not logged out when the
item level hold form is given to the user, and the autologout choice is
preserved in the item level hold form so that the log out can occur
properly after the item level hold is placed. #PK-1035