PK-407 Mobile Account & Login/out Button Styling & Templates
http://marmot.myjetbrains.com/youtrack/issue/PK-407

Adds Account icon & Logout icon in upper righthand of the screen on
mobile views only (using media rules) when user is logged in. Adds a
Login icon in the same position on mobile views only when the user is
not logged in.

The Account icon is a hyperlink to the anchor attached to the account
menu "sidebar" that is at the bottom in small  viewports.

The Login & Logout icons initiate standard log-in or log out behaviour.
#PK-407