PK-407  Mobile Menu
http://marmot.myjetbrains.com/youtrack/issue/PK-407

On initial page load with the top of the page shown in the (mobile)
browser, the menu is embedded with the bread crumb row. This avoids
interfering with system messages and the google translate buttons.  When
a user scrolls down, as the menu would go out of view, the menu affixes
to the top right corner. (Using a jquery function to toggle off & on a
css class triggered by user scrolling.) #PK-407