PK-603 Flatirons Hold Cancellation Issue
https://marmot.myjetbrains.com/youtrack/issue/PK-603

The ability to cancel required a pause between the curl login and curl
call to cancel in the same fashion we had to fix renewals recently.

The debugging item also revealed that we weren't handling failure
detection (of the hold cancel) correctly for some items, or the
retrieving of the item title for success/failure messages.  The has been
fixed for the cases that this item represents. #PK-603