PK-1156 Sidebar Fines for account(s)
http://marmot.myjetbrains.com/youtrack/issue/PK-1156

Pluralized when linked accounts are set.
Also changed references to the template variable 'profile' to 'user' as
the 'profile' variable changes when a linked account page is view for a
linked account (ie Reading History). So the sidebar now refers to the
main user or totals of all linked accounts. #PK-1156