PK-1007 Linked Accounts Hold placement.
http://marmot.myjetbrains.com/youtrack/issue/PK-1007

On the place hold form, if the user chooses a pick-up location that is
owned by more than one linked account , an additional drop down appears
allowing the user to choose the account to place a hold  with.
Previously, some pickup locations could appear multiple times in the
pickup select list.  Now duplicates are removed; and the location with
the higher priority sorting key is preserved.

Behind the scenes, the form supplies the account id to place the hold
with for any location. Greatly simplifying the server side processing of
the place hold request.  Should a selected account not  be provided in
the ajax request to submit the hold, the previous method of determining
the account to use will be used.

The success message now uses the account the hold was placed with so
that info about notification preferences is correct. #PK-1007