PK-1176 url parameter opac, PK-1284 Main Branch Location
http://marmot.myjetbrains.com/youtrack/issue/PK-1176
http://marmot.myjetbrains.com/youtrack/issue/PK-1284

Adding to the URL opac=1 or opac=true or opac=on will set the status of
the requesting machine as a opac.

This status turns on the auto log out functionality.

Added isMainBranch to Location table. Default value is false. For the
Library hours & locations display, Main branches will be listed first.
Also hide address & map if no address is present for selected location.
#PK-1176 #PK-1284