PK-1321  Easier to understand options to confirm when canceling hold
http://marmot.myjetbrains.com/youtrack/issue/PK-1321

Modal Box prompt
now employed to ask user to confirm cancelling hold. An
AJAX call is
made so that any custom translations can be used for the
confirmation
pop-up.  Translatable items are 'Cancel Hold' label,
'Confirm Cancel
Hold' and the question "Are you sure you want to cancel
this hold?"
#PK-1321