PK-1339  Pending Holds sorted alphabetically by title.
http://marmot.myjetbrains.com/youtrack/issue/PK-1339

Pending Holds are now sorted.  If the sort-title is available, that is
used over the title, otherwise the title is used to sort by.  (Not all
hold items have a sort title.)  It now uses a sort function which
preserves the array-key, which has item & user information. #PK-1339
state fixed