PK-1474 Prevent main content overflow with right hand side bar
http://marmot.myjetbrains.com/youtrack/issue/PK-1474

Scrollbar is put onto the main content div when the right side bar is
used to prevent the main content for spilling into the right hand side
bar.  Added in markup rather than style sheets so that it comes into
play only when the right hand side bar is turned on. #PK-1474