PK-410 Show/Hide Covers
http://marmot.myjetbrains.com/youtrack/issue/PK-410

Show or Hide Book Covers on MyAccount pages based on a device-stored
setting. Users check/uncheck a Hide Covers button. The choice is stored
in the browser's local storage.  When the setting is changed, it is
passed back to the server through the URL parameter showCovers and
stored for the user's session.  The CheckedOut, Holds & Reading History
templates will then show or hide covers based on the template switch
$showCovers  #PK-1612 #Pk-278 #PK-1611 #PK-1610