PK-1640 Hide Covers in Search Results.
http://marmot.myjetbrains.com/youtrack/issue/PK-1640

Covers switch is on the same line as the toggle display mode buttons.
When in covers display mode the Hide Covers switch is hidden.
Currently this uses the same switch as MyAccount Pages.

If the browser has a stored setting for covers, and the user in not
logged in, and several other conditions, the stored setting is added to
the search form so that the first search results are displayed matching
the setting. #PK-1640