ARL-190  MyAccount Mini Menu
http://marmot.myjetbrains.com/youtrack/issue/ARL-190

A Mini Menu added to MyAccount Pages that is shown at the top of
MyAccount pages when a menu option has not been selected.  When a menu
option is selected, the mini menu is hidden.  It is placed above the
available holds notice but below the webnote.

Mobile menu anchor link have href attributes with the related section Id
so that the selected section is brought into view. #ARL-190