ARL-103 Refresh Account Info Button in pop-up when clicking Ecommerce Link
https://marmot.myjetbrains.com/youtrack/issue/ARL-103

When a user clicks the Ecommerce link to pay fines, while the outside
link is being opened in a new tab or window, a modal pop-up loads for
the user so that on completion of payment process the user can click the
Refresh Account Info button.

Clicking the button will go to the Fines page, updating the user's
account information. This uses the url parameter 'reload', so all linked
accounts will be refreshed. (Account information stored in memcache will
be updated.) #ARL-103