Record Grouping updates #PK-1728
Update so we can just explode MARC records to make it easier to copy
settings from test/production systems.
Always add primary identifier to a work.  We were having cases where a
record was not part of an export for various reasons.  After it
reappeared the record did not get regrouped.