PK-237 Prevent Login for Expired Cards
http://marmot.myjetbrains.com/youtrack/issue/PK-237

Added Configuration switch to Library Admin preventExpiredCardLogin that
when enabled will return an error message on login that a card is
expired.

The Millennium Driver was updated to have better handling when the curl
login fails prior to other actions.

The full page Login templates were updated to present the error messages
in a cleaner way. #PK-237