PK-2165 Wake County Full Export Handling
http://marmot.myjetbrains.com/youtrack/issue/PK-2165

Previously another script called move_export.sh handled the full export
file and creating back-up copies. That is now integrated into the
full_update script and includes a file size check so that if the file is
not received or under a certain size, full regrouping and re-indexing
will not be run and the current index will be preserved. #PK-2165