PK-2227  Show Subject Types for Grouped Work View
https://marmot.myjetbrains.com/youtrack/issue/PK-2227

If a library has either LC subjects or Bisacs subjects chosen for
display but not  OCLC Fast Subjects or Other Subjects, the categories of
subjects will be diplayed in the grouped work view; otherwise all the
subjects will be shown in one list.

Each category is put in alphabetic order. #PK-2227