PK-2294  Update New York Times Lists
https://marmot.myjetbrains.com/youtrack/issue/PK-2294

Created a shell script that web-crawls through a list of Pika URLs that
update the New York Times User Lists within Pika.  Script is meant to be
set at a crontab job.

Error handling for a invalid list name was added to the ListAPI function
for creating the NY Times List.  #PK-2294