Archive fixes
On home page restrict materials to the archive based on namespace if the
library does not have sharing on and make the title match the spec.
#PK-2700
Update related collections for explore more to not show the section if
there is only 1 link (All Collections) and show as a text list if there
are more than 3.  #D-48